Adjust e2e tests
