+++ packages/common/src/invitationCode.ts
@@ -112,3 +112,21 @@ export const invitationCodeValid = (peerId: string, onionAddress: string): boole
   }
   return true
 }
+
+export const getInvitationPairs = (code: string) => {
+  /**
+   * @param code <peerId1>=<address1>&<peerId2>=<address2>
+   */
+  const pairs = code.split('&')
+  const codes: InvitationPair[] = []
+  for (const pair of pairs) {
+    const [peerId, address] = pair.split('=')
+    if (!peerId || !address) continue
+    if (!invitationCodeValid(peerId, address)) continue
+    codes.push({
+      peerId: peerId,
+      onionAddress: address,
+    })
+  }
+  return codes
+}
